+++ packages/extension-chakra-store-locator/src/pages/store-locator/index.tsx
@@ -7,6 +7,7 @@
 
 import React from 'react'
 import {Box, Container} from '@chakra-ui/react'
+
 import {StoreLocatorContent} from '../../components/content'
 
 const StoreLocatorPage = () => {