+++ packages/pwa-kit-extension-sdk/src/react/components/index.ts
@@ -1,8 +1,9 @@
 /*
- * Copyright (c) 2024, salesforce.com, inc.
+ * Copyright (c) 2025, salesforce.com, inc.
  * All rights reserved.
  * SPDX-License-Identifier: BSD-3-Clause
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 
 export {default as withApplicationExtensions} from './withApplicationExtensions'
+export {default as withApplicationExtensionStore} from './withApplicationExtensionStore'