+++ packages/pwa-kit-extension-sdk/src/react/components/withApplicationExtensions.tsx
@@ -45,12 +45,16 @@ const withApplicationExtensions = <C,>(
     WrappedComponent: React.ComponentType<C>,
     options: withApplicationExtensionsOptions
 ) => {
-    const hocs: GenericHocType<C>[] = options.applicationExtensions
+    const {applicationExtensions} = options
+
+    // Get all application extension higher-order components (HOCs)
+    const hocs: GenericHocType<C>[] = applicationExtensions
         .filter((applicationExtension) => applicationExtension.isEnabled())
         // It's counterintuitive: reversing the list is necessary, as we build the React tree from innermost
         .reverse()
         .map((extension) => extension.extendApp.bind(extension) as GenericHocType<C>)
         .filter(Boolean)
+
     const withApplicationExtensionsProvider: GenericHocType<any> = (WrappedComponent) => {
         const WithApplicationExtensionsProvider = (props: any) => (
             <ApplicationExtensionsProvider extensions={options.applicationExtensions}>