+++ packages/pwa-kit-extension-sdk/CHANGELOG.md
@@ -1,3 +1,4 @@
 ## v4.0.0-extensibility-preview.2 (Dec 09, 2024)
+- Add global store with slice support for extensions (#2214)[https://github.com/SalesforceCommerceCloud/pwa-kit/pull/2214]
 - Add support for `.force_overrides` project file. (#2207)[https://github.com/SalesforceCommerceCloud/pwa-kit/pull/2207]
 - Initial release of Extensibility SDK. (#2099)[https://github.com/SalesforceCommerceCloud/pwa-kit/pull/2099]